requirements.txt

wandb
streamlit
opencv-python
matplotlib
scikit-learn
pyyaml
pandas
